prep 4.0.6
removed debug messages
changes eeprom write size to 2 bytes
fixed eeprom writes
give distributed reference clock enough time
calculate sync manager start addresses if not set by user

+++ libethercat/include/libethercat/eeprom.h
@@ -180,7 +180,7 @@ int ec_eepromread(struct ec *pec, uint16_t slave,
  * \return 0 on success
  */
 int ec_eepromwrite(struct ec *pec, uint16_t slave, 
-        uint32_t eepadr, uint32_t *data);
+        uint32_t eepadr, uint16_t *data);
 
 //! read a burst of eeprom
 /*!
